requirements-txt.txt
torch
diffusers
Pillow
pyyaml
flask
